requirements.txt
psycopg2==2.9.10
python-dotenv==1.0.1
types-psycopg2==2.9.21.20241019
